+++ packages/bluesky/lib/src/entities/embed_view_record.freezed.dart
@@ -175,9 +175,10 @@ class _$_EmbedViewRecord implements _EmbedViewRecord {
 
 abstract class _EmbedViewRecord implements EmbedViewRecord {
   const factory _EmbedViewRecord(
-      {@JsonKey(name: '\$type') final String? type,
+      {@JsonKey(name: '\$type')
+          final String? type,
       @EmbedViewRecordViewConverter()
-      required final EmbedViewRecordView record}) = _$_EmbedViewRecord;
+          required final EmbedViewRecordView record}) = _$_EmbedViewRecord;
 
   factory _EmbedViewRecord.fromJson(Map<String, dynamic> json) =
       _$_EmbedViewRecord.fromJson;